+++ packages/tailor-ui/.size-snapshot.json
@@ -1,7 +1,7 @@
 {
   "lib/index.cjs.js": {
-    "bundled": 413438,
-    "minified": 287131,
-    "gzipped": 131523
+    "bundled": 414442,
+    "minified": 287570,
+    "gzipped": 131902
   }
 }
