+++ packages/tailor-ui/src/Breadcrumb/index.ts
@@ -1,0 +1,1 @@
+export * from './Breadcrumb';
